move recompute logic to standalone file
